+++ controllers/PacienteController.js
@@ -2,7 +2,7 @@ const db = require('../models')
 
 const Paciente = db.paciente
 
-const addPaciente = async (req, res ) => {
+const addPaciente = async (req, res) => {
 
   const paciente = await Paciente.create({
     nombre: req.body.paciente.nombre,
@@ -11,9 +11,9 @@ const addPaciente = async (req, res ) => {
     observacion: req.body.paciente.observacion ?? '',
     sexoId: req.body.paciente.sexo
   })
-  if(req.body.full){
+  if (req.body.full) {
     return paciente
-  }else{
+  } else {
     res.status(200).send(paciente)
   }
 }
@@ -26,42 +26,58 @@ const getPacientes = async (req, res) => {
 
 const getPacienteById = async (req, res) => {
 
-  let id = req.params.id
-  let paciente = await Paciente.findByPk(id)
-  res.status(200).send(paciente)
+  let paciente = await Paciente.findByPk(req.params.id)
+  if(paciente){
+    res.status(200).send(paciente)
+  }else{
+    res.status(404).send("No existe ningún paciente con id = " + req.params.id)
+  }
 }
 
 const deletePaciente = async (req, res) => {
 
-  let id = req.params.id
-  try {
-    let lesiones = await db.lesion.findAll({ where: { pacienteId: id } })
-    for (let lesion of lesiones) {
-      await db.consulta.destroy({
-        where: { lesionId: lesion.id }
+  const paciente = await Paciente.findByPk(req.params.id)
+  if (paciente) {
+    try {
+      let lesiones = await db.lesion.findAll({ where: { pacienteId: paciente.id } })
+      for (let lesion of lesiones) {
+        await db.consulta.destroy({
+          where: { lesionId: lesion.id }
+        })
+      }
+      await db.lesion.destroy({
+        where: { pacienteId: paciente.id },
       })
+      await Paciente.destroy({
+        where: { id: paciente.id }
+      })
+      res.status(200).send('Paciente borrado con todos sus registros')
+    } catch (error) {
+      res.status(500).send("Error: " + error.message)
     }
-    await db.lesion.destroy({
-      where: { pacienteId: id },
-    })
-    await Paciente.destroy({
-      where: { id: id }
-    })
-    res.status(200).send('Paciente borrado con todos sus registros')
-  } catch (error) {
-    res.status(500).send("Error: " + error.message)
+  } else {
+    res.status(404).send("No existe ningún paciente con id = " + req.params.id)
   }
 }
 
 const updatePaciente = async (req, res) => {
-
-  //! Te tira que todo bien cuando no existe la fila elegida para actualizar
-  if (Object.keys(req.body).every(el => Object.keys(Paciente.rawAttributes).includes(el))) {
-    await Paciente.update(req.body, { where: { id: req.params.id } })
-    res.status(200).send("Datos actualizados correctamente")
-  } else {
-    res.status(400).send("Datos incorrectos para la actualización")
+  
+  const paciente = await Paciente.findByPk(req.params.id)
+  if(paciente){
+    try {
+      if (Object.keys(req.body).every(el => Object.keys(Paciente.rawAttributes).includes(el))) {
+        await Paciente.update(req.body, { where: { id: req.params.id } })
+        res.status(200).send("Datos actualizados correctamente")
+      } else {
+        res.status(400).send("Datos incorrectos para la actualización")
+      }
+    } catch (error) {
+      res.status(500).send("Error: " + error.message)
+    }
+  }else{
+    res.status(404).send("No existe ningún paciente con id = " + req.params.id)
   }
+
 }
 
 module.exports = {
